Start migrating to MVVM
