+++ MPF/App.xaml
@@ -1,9 +1,13 @@
 ﻿<Application
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:MPF"
-             xmlns:Themes="clr-namespace:Microsoft.Windows.Themes;assembly=PresentationFramework.Aero" x:Class="MPF.App"
-             StartupUri="Windows\MainWindow.xaml">
+             xmlns:Themes="clr-namespace:Microsoft.Windows.Themes;assembly=PresentationFramework.Aero"
+             xmlns:windows="clr-namespace:MPF.Windows"
+             x:Class="MPF.App">
+    <Application.MainWindow>
+        <windows:MainWindow Visibility="Visible"/>
+    </Application.MainWindow>
+    
     <Application.Resources>
 
         <!-- Button -->
@@ -53,7 +57,7 @@
                 </Setter.Value>
             </Setter>
         </Style>
-        
+
         <!-- ComboBox -->
         <Style x:Key="FocusVisual">
             <Setter Property="Control.Template">
@@ -569,7 +573,7 @@
                 </Setter.Value>
             </Setter>
         </Style>
-        
+
         <!-- TabItem -->
         <LinearGradientBrush x:Key="TabItem.Static.Background" EndPoint="0,1" StartPoint="0,0">
             <GradientStop Color="#F0F0F0" Offset="0.0"/>
