+++ MPF/App.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows;
+using MPF.Windows;
 
 namespace MPF
 {
@@ -7,5 +8,19 @@ namespace MPF
     /// </summary>
     public partial class App : Application
     {
+        /// <summary>
+        /// Static application instance for reference
+        /// </summary>
+        private static App _appInstance;
+
+        /// <summary>
+        /// Read-only access to the current main window
+        /// </summary>
+        public static MainWindow Instance => _appInstance.MainWindow as MainWindow;
+
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        public App() => _appInstance = this;
     }
 }
