+++ MPF/UserControls/UserInput.xaml
@@ -2,19 +2,27 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:MPF.UserControls"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800" x:Name="userInput">
     <Grid>
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="1*" />
             <ColumnDefinition Width="1.25*" />
         </Grid.ColumnDefinitions>
+
+        <Label Grid.Row="0" Grid.Column="0"
+               VerticalAlignment="Center"
+               HorizontalAlignment="Right"
+               Content="{Binding ElementName=userInput, Path=Label}" />
         
-        <Label Grid.Row="0" Grid.Column="0" VerticalAlignment="Center" HorizontalAlignment="Right" Content="{Binding Label}" />
-        <TextBox Grid.Row="0" Grid.Column="1" Height="{Binding TextHeight}" HorizontalAlignment="Stretch" Text="{Binding Text}"
-                 AcceptsTab="{Binding Tab}" AcceptsReturn="{Binding Enter}" TextWrapping="{Binding TextWrapping}"
-                 VerticalContentAlignment="{Binding VerticalContentAlignmentValue}" VerticalScrollBarVisibility="{Binding ScrollBarVisibility}"/>
+        <TextBox Grid.Row="0" Grid.Column="1"
+                 Height="{Binding ElementName=userInput, Path=TextHeight}"
+                 HorizontalAlignment="Stretch"
+                 Text="{Binding ElementName=userInput, Path=Text}"
+                 AcceptsTab="{Binding ElementName=userInput, Path=Tab}"
+                 AcceptsReturn="{Binding ElementName=userInput, Path=Enter}"
+                 TextWrapping="{Binding ElementName=userInput, Path=TextWrapping}"
+                 VerticalContentAlignment="{Binding ElementName=userInput, Path=VerticalContentAlignmentValue}"
+                 VerticalScrollBarVisibility="{Binding ElementName=userInput, Path=ScrollBarVisibility}"/>
     </Grid>
 </UserControl>
