+++ MPF/Windows/OptionsWindow.xaml
@@ -1,10 +1,10 @@
-﻿<Window x:Class="MPF.Windows.OptionsWindow"
+﻿<windows:WindowBase x:Class="MPF.Windows.OptionsWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:MPF"
-        xmlns:controls="clr-namespace:MPF.UserControls"
+        xmlns:windows="clr-namespace:MPF.Windows"
         mc:Ignorable="d"
         Title="Options" Width="515.132" WindowStyle="None"
         WindowStartupLocation="CenterOwner" ResizeMode="CanMinimize" SizeToContent="Height"
@@ -92,17 +92,17 @@
                         <Label Grid.Row="0" Grid.Column="0" VerticalAlignment="Center" HorizontalAlignment="Right" Content="Aaru Path" />
                         <TextBox x:Name="AaruPathTextBox" Grid.Row="0" Grid.Column="1" Height="22" HorizontalAlignment="Stretch" Text="{Binding Options.AaruPath}" VerticalContentAlignment="Center" />
                         <Button x:Name="AaruPathButton" Grid.Row="0" Grid.Column="2" Height="22" Width="22" Content="..."
-                                Click="BrowseForPathClick" Style="{DynamicResource CustomButtonStyle}" />
+                                Style="{DynamicResource CustomButtonStyle}" />
 
                         <Label Grid.Row="1" Grid.Column="0" VerticalAlignment="Center" HorizontalAlignment="Right" Content="DiscImageCreator Path" />
                         <TextBox x:Name="DiscImageCreatorPathTextBox" Grid.Row="1" Grid.Column="1" Height="22" HorizontalAlignment="Stretch" Text="{Binding Options.DiscImageCreatorPath}" VerticalContentAlignment="Center" />
                         <Button x:Name="DiscImageCreatorPathButton" Grid.Row="1" Grid.Column="2" Height="22" Width="22" Content="..."
-                                Click="BrowseForPathClick" Style="{DynamicResource CustomButtonStyle}" />
+                                Style="{DynamicResource CustomButtonStyle}" />
 
                         <Label Grid.Row="2" Grid.Column="0" VerticalAlignment="Center" HorizontalAlignment="Right" Content="DD Path" />
                         <TextBox x:Name="DDPathTextBox" Grid.Row="2" Grid.Column="1" Height="22" HorizontalAlignment="Stretch" Text="{Binding Options.DDPath}" VerticalContentAlignment="Center" />
                         <Button x:Name="DDPathButton" Grid.Row="2" Grid.Column="2" Height="22" Width="22" Content="..."
-                                Click="BrowseForPathClick" Style="{DynamicResource CustomButtonStyle}" />
+                                Style="{DynamicResource CustomButtonStyle}" />
 
                         <Label Grid.Row="3" Grid.Column="0" VerticalAlignment="Center" HorizontalAlignment="Right" Content="Dumping Program" />
                         <ComboBox x:Name="InternalProgramComboBox" Grid.Row="3" Grid.Column="1" Height="22" HorizontalAlignment="Stretch"
@@ -111,7 +111,7 @@
                         <Label Grid.Row="4" Grid.Column="0" VerticalAlignment="Center" HorizontalAlignment="Right" Content="Default Output Path" />
                         <TextBox x:Name="DefaultOutputPathTextBox" Grid.Row="4" Grid.Column="1" Height="22" HorizontalAlignment="Stretch" Text="{Binding Options.DefaultOutputPath}" VerticalContentAlignment="Center" />
                         <Button x:Name="DefaultOutputPathButton" Grid.Row="4" Grid.Column="2" Height="22" Width="22" Content="..."
-                                Click="BrowseForPathClick" Style="{DynamicResource CustomButtonStyle}" />
+                                Style="{DynamicResource CustomButtonStyle}" />
                     </Grid>
                 </TabItem>
 
@@ -350,7 +350,7 @@
                                 <PasswordBox x:Name="RedumpPasswordBox" Height="22" HorizontalAlignment="Stretch" PasswordChar="*" />
 
                                 <Button x:Name="RedumpLoginTestButton" Height="22" Width="80" Content="Test Login"
-                                        Click="OnRedumpTestClick" Style="{DynamicResource CustomButtonStyle}" />
+                                        Style="{DynamicResource CustomButtonStyle}" />
                             </UniformGrid>
                         </GroupBox>
                     </StackPanel>
@@ -362,12 +362,12 @@
                 <UniformGrid Columns="4" Margin="5,5,5,5" Height="28">
                     <Label/> <!-- Empty label for padding -->
                     <Button Name="AcceptButton" Height="25" Width="80" IsDefault="True" Content="Accept"
-                            Click="OnAcceptClick" Style="{DynamicResource CustomButtonStyle}" />
+                            Style="{DynamicResource CustomButtonStyle}" />
                     <Button Name="CancelButton" Height="25" Width="80" IsCancel="True" Content="Cancel"
-                            Click="OnCancelClick" Style="{DynamicResource CustomButtonStyle}" />
+                            Style="{DynamicResource CustomButtonStyle}" />
                     <Label/> <!-- Empty label for padding -->
                 </UniformGrid>
             </GroupBox>
         </StackPanel>
     </Grid>
-</Window>
+</windows:WindowBase>
