+++ MPF/Windows/RingCodeGuideWindow.xaml
@@ -1,10 +1,11 @@
-﻿<Window x:Class="MPF.Windows.RingCodeGuideWindow"
+﻿<windows:WindowBase x:Class="MPF.Windows.RingCodeGuideWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:MPF"
         xmlns:controls="clr-namespace:MPF.UserControls"
+        xmlns:windows="clr-namespace:MPF.Windows"
         mc:Ignorable="d"
         Title="Ring Code Guide" Width="500" WindowStyle="None"
         WindowStartupLocation="CenterOwner" ResizeMode="CanMinimize" SizeToContent="Height"
@@ -100,4 +101,4 @@
             </Grid>
         </StackPanel>
     </Grid>
-</Window>
+</windows:WindowBase>
